get challenges as well as skills for ~available
